IdentifierFactory 분리
- ViewRepresentable 프로토콜이 Identifier를 포함하지 않기 떄문에 Identifier를 생성하는 Factory 분리
