+++ DrawingApp/DrawingApp/Model/Protocol/IdentifierFactory.swift
@@ -1,0 +1,12 @@
+//
+//  IdentifierFactory.swift
+//  DrawingApp
+//
+//  Created by Noah on 2023/03/29.
+//
+
+import Foundation
+
+protocol IdentifierFactory {
+    func makeIdentifier() -> Identifier
+}
